requirements.txt
django
djangorestframework
markdown
django-environ
django-filter
mysqlclient
pyyaml 
ua-parser 
user-agents 
django-user-agents
python-memcached
mailchimp-marketing
dj-database-url
gunicorn
psycopg2
whitenoise
django-cors-headers
sendgrid